Minor improvements to the core
Every message used to have size 256, which would usually lead to a very
large message, with a load of 0'ed bytes.

We now truncate that to whatever was read from the connection.

We also had a possibly leaky goroutine, which is now covered from the
point of view of the WaitGroup.