Moving to interface-based approach
The environment can now be configured with a Backend.

The Backend dictates how to act on specific data collections (one for
each MongoDB collection) and it also introduces a MongoDB Backend.

This backend can be easily swapped if, in the future, we decide to
support a SQL database, or even a cache, for example.

The interface contains CRUD methods for each collection. This simplified
logic in the handlers and enables a much easier testing.

Further commits should address tests for each handler, although
achieving 100% coverage is _not_ (as of yet) the goal of this.